pyproject.toml
[project]
name = "kubeagent"
version = "0.1.0"
description = "A natural language CLI agent for Kubernetes cluster management"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "KubeAgent Contributors" },
]
keywords = ["kubernetes", "k8s", "cli", "agent", "ai"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
]

[project.scripts]
kubeagent = "kubeagent.cli.main:cli"

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kubeagent"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["kubeagent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
